extra cleanup

+++ src/main/java/com/sudolev/interiors/block/seat/ChairBlock.java
@@ -155,15 +155,14 @@ public class ChairBlock extends DirectionalSeatBlock implements ProperWaterlogge
             }
 
             Vec3 vec3 = entity.getDeltaMovement();
-            if (vec3.y < 0.0D) {
-                double d0 = entity instanceof LivingEntity ? 1.0D : 0.8D;
-                entity.setDeltaMovement(vec3.x, -vec3.y * (double) 0.66F * d0, vec3.z);
+            if (vec3.y < 0) {
+                double d0 = entity instanceof LivingEntity ? 1 : 0.8;
+                entity.setDeltaMovement(vec3.x, -vec3.y * 0.66 * d0, vec3.z);
             }
 
             return;
         }
-        if (reader.getBlockState(pos)
-                .getBlock() != this)
+        if (reader.getBlockState(pos).getBlock() != this)
             return;
         sitDown(entity.level(), pos, entity);
     }
